Implemented custom non-blocking live reload server that enables running the app while a run happens in the background
