+++ pyproject.toml
@@ -19,7 +19,6 @@ sql-formatter = "^0.6.2"
 click = "^8.1.3"
 pandas = "^1.5.3"
 Flask = "^2.2.3"
-livereload = "^2.6.3"
 tornado = "< 6.3"
 numpy = "=1.26.0"
 psycopg2-binary = "^2.9.5"
@@ -46,6 +45,7 @@ snowflake-sqlalchemy = "^1.6.1"
 sqlalchemy-bigquery = "^1.12.0"
 duckdb-engine = "^0.14.0"
 playwright = "^1.50.0"
+flask-socketio = "^5.5.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.0.0"
